Assignment 1 commenced
Modified .gitignore to ignore the datasets directory to avoid transfer
of huge files.
Modified README to reflect the same.
Completed till question 7 of the assignment 1 on the Jupyter notebook
Please do download the MNIST dataset and then add it to the datasets
folder, or just modify the code on my system and then run it on the
server directly.