Complete 8, 9 derivation. Complete relup and relu
Completed typing in LaTex the derivation for questions 8 and 9.
Derived question 10 on paper, and then implemented it.
Tested my implementation of the directional derivative of the softmax()
function using 1st principle approximations of the derivative.
Implemented ReLU and computed it's directional derivative on paper using
Jacobians.
Tested the implementation of relup() using 1st principle approximations
of the derivative.

TODO -
 - Complete typing out / scanning and inserting into the Jupyter
   notebook the derivative of the softmax() function for question 10.
 - Complete typing out / scannin and inserting into the Jupyter notebook
   the derivative of the relup() function for question 12.
 - Complete the rest of the assignment starting from backpropagation