Complete backprop for 20 iterations
Completed questions 13 through 17.
Initialized the network using He and Xavier initialization method for
the 1st and 2nd layer respectively.
Completed the forward_prop_shallow() function to compute the output of
the network
Completed eval_loss() and eval_perfs() to compute the loss and the
performance of the network.
Wrote code to compute the output of a single backpropagation operation.
Used that code to perform backpropagation on the entire network.
Trained the network for 2, 5, and 20 iterations to test if it worked

TODO - Complete question 18 by running the network for 100 iterations
Complete the questions 19 through 21.
Complete the proofs for questions 10 and 11 to complete the assignment