Finish questions 7 through 10
Completed the assignment 3 till question 10.
Inspected the abstract class NeuralNetwork, and defined my own abstract class NNClassifier on top of that.
Loaded the pre-trained VGG16 network defined with BatchNorm and AdaptiveAveragePool.
Inherited from NNClassifier and defined a network that implemented the forward() function, thus making it a concrete class.
Also, froze all the VGG16 pre-trained layers, except a new final layer that I shall be training by creating a new Linear layer that is consistent with the number of classes that I'm trying to classify into.
On printing the network and inspecting its learnable parameters, I find that the only learnable parameters are for the newly defined final layer.